docs: updating docs to include auth changes
