+++ Sources/Networking/Authentication/AuthenticationProvider.swift
@@ -1,3 +1,4 @@
+/// Protocol for passing a bearer token for authentication
 public protocol AuthenticationProvider {
     var bearerToken: String { get async throws }
 }
